requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
diskcache==5.6.3
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.3.2
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==25.0
pfzy==0.3.4
pillow==11.2.1
prompt_toolkit==3.0.51
pyclipper==1.3.0.post6
PyMuPDF==1.25.5
python-bidi==0.6.6
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.7.1
shapely==2.1.0
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
